requirements.txt
fastapi==0.115.11
python-dotenv==1.0.1
msal==1.32.3
python-socketio==5.13.0
uvicorn==0.34.0
opencv-python==4.11.0.86
pytesseract==0.3.13
pdfminer==20191125
pdfminer.six==20250506
openai==1.78.1
python-multipart==0.0.20
selenium==4.32.0
deepgram-sdk==4.1.0
pydantic[email]>=2.11.4
